Use int instead of long

+++ src/main/java/br/com/waterclock/api/repository/UserRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
     User findByEmail(String email);
-    User findById(Long id);
+    User findById(int id);
 }
